@@ -19,10 +19,10 @@
  FOR A PARTICULAR PURPOSE.  See the license for more details.
 */
 
+#include <ql/exercise.hpp>
 #include <ql/instruments/barrieroption.hpp>
 #include <ql/instruments/impliedvolatility.hpp>
 #include <ql/pricingengines/barrier/analyticbarrierengine.hpp>
-#include <ql/exercise.hpp>
 #include <memory>
 
 namespace QuantLib {
@@ -67,8 +67,8 @@ namespace QuantLib {
         std::unique_ptr<PricingEngine> engine;
         switch (exercise_->type()) {
           case Exercise::European:
-            engine.reset(new AnalyticBarrierEngine(newProcess));
-            break;
+              engine = std::make_unique<AnalyticBarrierEngine>(newProcess);
+              break;
           case Exercise::American:
           case Exercise::Bermudan:
             QL_FAIL("engine not available for non-European barrier option");
